requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
passlib[bcrypt]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
